update DB and scheme

+++ application/controllers/Migration.php
@@ -56,13 +56,18 @@ class Migration extends CI_Controller {
 
 	}
 
+	/**
+	 * drop all tables
+	 * 
+	 * @return null
+	 */
     protected function drop_all() {
 
         $data = $this->mc->showAllTables();
 
         // drop database table with prefix setting
         foreach ($data AS $key => $r){
-            $table = str_replace($_ENV['DB_PREFIX'], '', $key);
+            $table = str_replace(getenv('DB_PREFIX'), '', $key);
 
             // drop data
             $this->schema->drop_table($table);
@@ -70,6 +75,11 @@ class Migration extends CI_Controller {
         
     }
 
+	/**
+	 * drop only one table
+	 * 
+	 * @return null
+	 */
     protected function drop_one($name) {
         $this->schema->drop_table($name);
 	}
@@ -81,12 +91,8 @@ class Migration extends CI_Controller {
 	 * @return void
 	 */
 	public function create( $dbname = null ){
-
 		// create migration
-		if( $dbname !== null ){
-			$this->migrate->create_file($dbname);
-		}
-
+		if( $dbname !== null ) $this->migrate->create_file($dbname);
 	}
 
 	/**
@@ -97,49 +103,42 @@ class Migration extends CI_Controller {
 	 * @return void
 	 */
 	public function install($action = null, $tablename = null){
-        
-        if($action == 'seed'){
+		switch($action) {
+			case 'seed':
+					if($tablename == null)$this->drop_all();
+					else $this->drop_one($tablename);
+					$this->migrate($tablename);
+					$this->seed($tablename);
+				break;
+			default: 
+				if( $action === null ) $this->migrate();
+				else {
+					$tablename = strtolower($action);
+					$this->migrate($tablename);
+				}
+				break;
 
-			if($tablename == null){
-				$this->drop_all();
-			} else {
-				$this->drop_one($tablename);
-			}
-
-			$this->migrate($tablename);
-			$this->seed($tablename);
-
-		} else {
-
-			if( $action != null ){
-				$tablename = strtolower($action);
-				$this->migrate($tablename);
-
-			} else {
-				$this->migrate();
-			}
 		}
+		//
     }
 	
-    // ======= MIGRATE
+	/**
+	 * migration for migrate table
+	 * 
+	 * @param string $tablename
+	 * @return null
+	 */
 	protected function migrate( $tablename = null ){
-
 		if($tablename!=null){
-
 			$filemigration = 'Table_' . strtolower($tablename);
 			if( is_file($this->locatemigration.$filemigration.'.php') ){
-
 				$this->migrate_action($tablename);
-
 			} else {
 				echo "\n".'Sorry,, File '.$filemigration.'.php is not already exist ^_^'."\n\n";exit;
 			}
-
 		} else {
-
 			// install all seeder here
 			if( is_dir($this->locatemigration) ){
-
 				$migrationsdir = scandir($this->locatemigration);
 				
 				foreach ($migrationsdir as $thefile) {
@@ -152,12 +151,19 @@ class Migration extends CI_Controller {
 						$this->migrate_action($tablename);
 					}
 				}
-
+				//
 			}
+			//
 		}
-
+		//
 	}
 	
+	/**
+	 * migration action for migration
+	 * 
+	 * @param string $tablename
+	 * @return null
+	 */
 	protected function migrate_action($tablename){
 
 		if($tablename!=null){
@@ -180,26 +186,25 @@ class Migration extends CI_Controller {
 
 	}
 
-    // ======= SEED
+	/**
+	 * seed for insert data
+	 * 
+	 * @param string $tablename
+	 * 
+	 * @return null
+	 */
     protected function seed( $tablename = null ) {
-
 		if($tablename !=null){
-			
 			$filemigration = 'Table_' . strtolower($tablename);
 			if( is_file($this->locatemigration.$filemigration.'.php') ){
-
 				$this->seeder_action($tablename);
-
 			} else {
 				echo "\n".'Sorry,, File '.$filemigration.'.php is not already exist ^_^'."\n\n";exit;
 			}
-
 		} else {
 			// install all seeder here
 			if( is_dir($this->locatemigration) ){
-
 				$migrationsdir = scandir($this->locatemigration);
-				
 				foreach ($migrationsdir as $thefile) {
 					if ($thefile != "." && $thefile != ".."){
 						if( $thefile =='index.html' ){ continue; }
@@ -216,9 +221,14 @@ class Migration extends CI_Controller {
 		}
 	}
 
+	/**
+	 * seeder action
+	 * 
+	 * @param string @tablename
+	 * @return null
+	 */
 	protected function seeder_action($tablename = null){
-
-		if($tablename!=null){
+		if($tablename != null){
 			$tablename = strtolower($tablename);
 			$filemigration = 'Table_' . $tablename;
 	
@@ -237,34 +247,32 @@ class Migration extends CI_Controller {
 					}
 				}
 			}
-
+			//
 		}
-
+		//
 	}
 
-
-	// ======= REMOVE
+	/**
+	 * remove table
+	 * 
+	 * @param string $tablename
+	 * @return null
+	 */
 	public function remove($tablename = null){
-
 		if($tablename != null){
 			$tablename = strtolower($tablename);
 			$filemigration = 'Table_' . $tablename;
 			if( is_file($this->locatemigration.$filemigration.'.php') ){
-
 				// unlink file
 				@unlink( $this->locatemigration.$filemigration.'.php' );
 
 				// drop table
 				$this->drop_one($tablename);
-
 			}
-
 		} else {
 			// install all seeder here
 			if( is_dir($this->locatemigration) ){
-
 				$migrationsdir = scandir($this->locatemigration);
-				
 				foreach ($migrationsdir as $thefile) {
 					if ($thefile != "." && $thefile != ".."){
 						if( $thefile =='index.html' ){ continue; }
@@ -280,10 +288,88 @@ class Migration extends CI_Controller {
 						$this->drop_one($tablename);
 					}
 				}
+				//
+			}
+		}
+		//
+	}
+
+	/**
+	 * alter table for modify table
+	 * 
+	 * php index.php Migration alter ${table_name} ~> first init if column didnt't exists
+	 * php index.php Migration alter ${table_name} false ~> first init if column didnt't exists all file table migration
+	 * php index.php Migration alter ${table_name} refresh ~> if column is exists
+	 * php index.php Migration alter ${table_name} drop ~> only drop column without add column
+	 * 
+	 * @param string $tablename
+	 * @return null
+	 */
+	public function alter($tablename = null, $drop = false) {
+		if($tablename !== null && $tablename !== 'false') {
+			$filemigration = 'Table_' . strtolower($tablename);
+			// only drop without add column
+			if($drop === 'drop') $drop = 'drop';
+			else $drop = $drop === 'refresh' ? true : false;
+	
+			if(!is_null($tablename)) {
+				$file = $this->locatemigration . $filemigration . '.php';
+				if( is_file($file) ) $this->alter_action($tablename, $drop);
+				else echo "\n".'Sorry,, File '.$filemigration.'.php is not already exist ^_^'."\n\n";exit;
+			} else {
+				echo "\n".'Sorry,, File '.$filemigration.'.php is not already exist ^_^'."\n\n";exit;
+			}
+		} else {
+			// install all seeder here
+			if( is_dir($this->locatemigration) ){
+				$migrationsdir = scandir($this->locatemigration);
+				foreach ($migrationsdir as $thefile) {
+					if ($thefile != "." && $thefile != ".."){
+						if( $thefile =='index.html' ){ continue; }
+	
+						$filename = pathinfo($thefile, PATHINFO_FILENAME );
+
+						$tablename = str_replace("Table_", "", $filename);
+						$this->alter_action($tablename, $drop);
+					}
+				}
 
 			}
 		}
+	}
 
+	/**
+	 * alter action for modify table
+	 * 
+	 * @param string $tablename
+	 * @return null
+	 */
+	protected function alter_action($tablename = null, $drop = false) {
+		if($tablename != null){
+			$tablename = strtolower($tablename);
+			$filemigration = 'Table_' . $tablename;
+	
+			// load library
+			$this->load->library('migrations/'.$filemigration);
+	
+			$obj = $filemigration;
+			if( method_exists($obj, 'alter') ){	
+				$obj = strtolower($obj);
+				$fields = $this->$obj->alter($tablename);
+				$schema = $this->schema->create_table($tablename);
+				if($drop) {
+					foreach($fields AS $field => $value) {
+						$schema->drop_column(['table' => $tablename, 'field' => $field]);
+					}
+				}
+				
+				if($drop !== 'drop') $schema->add_column([ 'table' => $tablename, 'fields' => $fields ]);
+		
+				echo 'Successfully alter the table ' . $tablename . PHP_EOL;
+			}
+			//
+		}
+		//
 	}
     
 }
